+++ ElectricalCircuitSimulator/SimulationEngine/include/Inductor.h
@@ -9,7 +9,7 @@ namespace SimulationEngine {
 
         public:
 
-            Inductor(const int iNodeS, const int iNodeD, const double m_dInductance);
+            Inductor(const size_t iNodeS, const size_t iNodeD, const double m_dInductance);
 
             void initalize(Matrix& oConductanceMatrix, const double dTimeStep);
             void step(Matrix& oSourceVector); // Trapezoidal integration
